Move the only lookup test into its own test module
