+++ client_test/stub_test.py
@@ -9,7 +9,7 @@ from grpclib import GRPCError, Status
 
 import modal.app
 from modal import Stub
-from modal.aio import AioApp, AioQueue, AioStub, aio_lookup
+from modal.aio import AioApp, AioQueue, AioStub
 from modal.exception import InvalidError, NotFoundError
 from modal_proto import api_pb2
 from modal_test_support import module_1, module_2
@@ -38,24 +38,6 @@ async def test_attrs(servicer, aio_client):
         await app.q2.put("bar")
         assert await app.q1.get() == "foo"
         assert await app.q2.get() == "bar"
-
-
-@pytest.mark.asyncio
-async def test_persistent_object(servicer, aio_client):
-    stub_1 = AioStub()
-    stub_1["q_1"] = AioQueue()
-    await stub_1.deploy("my-queue", client=aio_client)
-
-    stub_2 = AioStub()
-    async with stub_2.run(client=aio_client) as app_2:
-        assert isinstance(app_2, AioApp)
-
-        q_3 = await aio_lookup("my-queue", client=aio_client)
-        # assert isinstance(q_3, AioQueue)  # TODO(erikbern): it's a AioQueueHandler
-        assert q_3.object_id == "qu-1"
-
-        with pytest.raises(NotFoundError):
-            await aio_lookup("bazbazbaz", client=aio_client)
 
 
 def square(x):
